feat: add new services, improve existing APIs, and fix comment N+1
- Add analytics proof-of-work and token services for bot mitigation
- Add user cache service for faster lookups
- Add batch comment replies loading (fixes N+1 query)
- Improve article service with circuit breaker integration
- Add role validation via UserRole enum in role upgrade requests
- Update controllers with consistent error handling
- Add Cloudflare email routing and LinkedIn DMA portability services
- Remove deprecated AdminImageController (merged into MediaService)
- Update repositories with new query methods

+++ src/main/java/dev/catananti/controller/EmailTemplatePreviewController.java
@@ -363,7 +363,7 @@ public class EmailTemplatePreviewController {
 
     /** Update a custom variable by ID. */
     @PutMapping("/custom-variables/{id}")
-    public Mono<Map<String, Object>> updateCustomVariable(@PathVariable int id, @RequestBody Map<String, String> body) {
+    public Mono<Map<String, Object>> updateCustomVariable(@PathVariable long id, @RequestBody Map<String, String> body) {
         String value = body.get("value");
         if (value == null) {
             return Mono.error(new ResponseStatusException(
@@ -377,7 +377,7 @@ public class EmailTemplatePreviewController {
 
     /** Delete a custom variable by ID. */
     @DeleteMapping("/custom-variables/{id}")
-    public Mono<Map<String, String>> deleteCustomVariable(@PathVariable int id) {
+    public Mono<Map<String, String>> deleteCustomVariable(@PathVariable long id) {
         return templateService.deleteCustomVariable(id)
             .map(deleted -> deleted
                 ? Map.of("message", "Variable deleted")
